pyproject.toml
[project]
name = "urban-prediction-mpi-index"
version = "0.1.0"
description = "Pipeline for preprocessing geospatial data and computing the Mazziotta-Pareto Index (MPI) for urban analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "geopandas>=0.14.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "h3>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mpi-preprocess = "src.preprocessing:main"
mpi-features = "src.feature_engineering:main"
mpi-pipeline = "src.pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
